Move collapsed bindable back to `BeatmapAttributesDisplay` for now
Better to have it only in one place that needs it, rather than have it
not work as expected when someone inherits
`ModFooterInformationDisplay`.
