+++ osu.Game/Overlays/Mods/ModFooterInformationDisplay.cs
@@ -2,12 +2,10 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Input.Events;
 using osu.Game.Graphics.UserInterface;
 using osuTK;
 using osuTK.Graphics;
@@ -20,8 +18,6 @@ namespace osu.Game.Overlays.Mods
         protected FillFlowContainer RightContent { get; private set; } = null!;
         protected Container Content { get; private set; } = null!;
 
-        public BindableBool Collapsed { get; } = new BindableBool();
-
         private Container innerContent = null!;
 
         protected Box MainBackground { get; private set; } = null!;
@@ -29,8 +25,6 @@ namespace osu.Game.Overlays.Mods
 
         [Resolved]
         private OverlayColourProvider colourProvider { get; set; } = null!;
-
-        private const float transition_duration = 250;
 
         [BackgroundDependencyLoader]
         private void load()
@@ -88,7 +82,6 @@ namespace osu.Game.Overlays.Mods
                             },
                             RightContent = new FillFlowContainer
                             {
-                                Alpha = 0,
                                 Origin = Anchor.CentreLeft,
                                 Anchor = Anchor.CentreLeft,
                                 AutoSizeAxes = Axes.X,
@@ -111,43 +104,6 @@ namespace osu.Game.Overlays.Mods
 
             Content.BorderColour = ColourInfo.GradientVertical(MainBackground.Colour, glowColour);
             innerContent.BorderColour = ColourInfo.GradientVertical(FrontBackground.Colour, glowColour);
-
-            Collapsed.BindValueChanged(_ =>
-            {
-                // Only start autosize animations on first collapse toggle. This avoids an ugly initial presentation.
-                startAnimating();
-                updateCollapsedState();
-            });
-
-            updateCollapsedState();
-        }
-
-        protected override bool OnHover(HoverEvent e)
-        {
-            startAnimating();
-            updateCollapsedState();
-            return true;
-        }
-
-        protected override void OnHoverLost(HoverLostEvent e)
-        {
-            updateCollapsedState();
-            base.OnHoverLost(e);
-        }
-
-        protected override bool OnMouseDown(MouseDownEvent e) => true;
-
-        protected override bool OnClick(ClickEvent e) => true;
-
-        private void startAnimating()
-        {
-            Content.AutoSizeEasing = Easing.OutQuint;
-            Content.AutoSizeDuration = transition_duration;
-        }
-
-        private void updateCollapsedState()
-        {
-            RightContent.FadeTo(Collapsed.Value && !IsHovered ? 0 : 1, transition_duration, Easing.OutQuint);
         }
     }
 }
